rgb-bridge/requirements.txt
Flask==3.0.3
Pillow==10.3.0
requests==2.31.0
gunicorn==22.0.0
